requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.4,<6.0
django-filter>=24.3
django-cors-headers>=4.6,<5.0
Pillow>=11.0,<12.0

